rag-system/backend/requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
async-timeout==4.0.3
aiosignal==1.4.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.14.2
billiard==4.2.2
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
cbor==1.0.0
celery==5.3.6
certifi==2025.10.5
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmake==4.1.2
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==4.3.0
decorator==5.2.1
diffusers==0.35.2
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docx2txt==0.8
durationpy==0.10
exceptiongroup==1.2.2
executing==2.2.1
faiss-cpu==1.8.0
fastapi==0.110.1
filelock==3.18.0
FlagEmbedding==1.3.5
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.42.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ijson==3.4.0.post0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
inscriptis==2.6.0
ipython==8.37.0
ipython_pygments_lexers==1.1.1
ir_datasets==0.5.11
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.0
kubernetes==34.1.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.2
langchain-text-splitters==0.3.2
langsmith==0.1.147
lit==18.1.8
lxml==6.0.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.10.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
networkx==3.4.2
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
ollama==0.3.1
onnxruntime==1.23.2
openai==1.40.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.10.7
overrides==7.7.0
packaging==24.2
pandas==2.3.1
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==10.4.0
pip==25.3
pluggy==1.6.0
posthog==3.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.11.0
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.24.12
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pycryptodome==3.21.0
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
redis==5.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==3.0.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.37.2
streamlit==1.50.0
sympy==1.14.0
tavily-python==0.5.0
tenacity==8.5.0
tesseract==0.1.3
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trec-car-tools==2.6
typer==0.20.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.3.0
uvicorn==0.30.1
uvloop==0.22.1
vine==5.1.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zlib-state==0.1.10
zstandard==0.23.0
